requirements.txt
# Core dependencies for classroom behavior detection
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0

# Database dependencies
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0

# MLflow for experiment tracking (preserved for model tracking)
mlflow>=2.0.0

# Desktop GUI
PySide6>=6.5.0

# Testing dependencies
hypothesis>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
